Refactored chat to chatbubbles
